fix(ui): reconcile composer rebase checks
