+++ ui/src/pages/new-session/new-session-page.ts
@@ -539,6 +539,7 @@ export class NewSessionPage extends OpenClawLightDomElement {
       this.place.worktree && !isWorktreeNameValid(this.place.worktreeName);
     const capabilities = this.submission.capabilities;
     const voiceControl = this.dictation.render(this.routeOwnerKey());
+    const dictationLocked = this.dictation.active;
     return html`
       <div class="new-session-page__draft" aria-busy=${String(this.submission.submitting)}>
         ${this.renderTargetBar()}
@@ -563,7 +564,7 @@ export class NewSessionPage extends OpenClawLightDomElement {
           agent: this.place.selectedAgent(),
           agentId: this.place.agentId,
           attachmentDraft: this.submission.attachmentDraft,
-          canSubmit: !this.submission.submitting && this.submission.canSubmit(),
+          canSubmit: !this.submission.submitting && !dictationLocked && this.submission.canSubmit(),
           submitDisabledReason: this.submission.submitDisabledReason(),
           blockedSubmitNotice: this.submission.blockedSubmitNotice(),
           dictationActive: this.dictation.active,
