+++ ui/src/styles/chat/composer-progress.css
@@ -143,6 +143,10 @@
   display: none;
 }
 
+.session-progress-card--composer[open] .session-progress-card__summary-indicator {
+  display: none;
+}
+
 .session-progress-card--composer[open] .session-progress-card__summary-count--collapsed {
   display: none;
 }
@@ -285,8 +289,7 @@
      the composer's rounded corners), so the highlight lives on the container to
      fill the whole shape; a summary-only fill leaves slivers of the resting
      background at those corners. */
-  .session-progress-card--composer:not([open]):hover,
-  .session-progress-card--composer[open] .session-progress-card__summary:hover {
+  .session-progress-card--composer:not([open]):hover {
     background: color-mix(
       in srgb,
       var(--text-strong) 5%,
